Create Fermat primality test
Fermat's little theorem (see also Euler's totient function) states, that for a prime number p and a coprime integer a the following equation holds:

ap−1≡1modp
In general this theorem doesn't hold for composite numbers.

This can be used to create a primality test. We pick an integer 2≤a≤p−2, and check if the equation holds or not. If it doesn't hold, e.g. ap−1≢1modp, we know that p cannot be a prime number. In this case we call the base a a Fermat witness for the compositeness of p.